requirements.txt
requests
pandas
PyJWT
PyYAML